feat(web): rework Routes as cards + own chain-editor page
Routes have no create/delete (a fixed, backend-defined set) so there
is no add page — RoutesList shows them as cards (name, strategy,
serving status), each opening its own /settings/routes/:name page for
reordering and adding chain entries.
